+++ lib/src/internal/layout/add_item.dart
@@ -1,5 +1,6 @@
 import 'package:docking/src/internal/layout/drop_item.dart';
 import 'package:docking/src/layout/docking_layout.dart';
+import 'package:docking/src/layout/drop_position.dart';
 import 'package:meta/meta.dart';
 
 /// Adds [DockingItem] to the layout.