+++ lib/src/internal/widgets/drop/content_wrapper.dart
@@ -1,5 +1,6 @@
 import 'package:docking/src/internal/widgets/drop/drop_anchor_widget.dart';
 import 'package:docking/src/layout/docking_layout.dart';
+import 'package:docking/src/layout/drop_position.dart';
 import 'package:flutter/material.dart';
 import 'package:meta/meta.dart';