+++ lib/src/layout/docking_area_type.dart
@@ -1,0 +1,2 @@
+/// Represents the [DockingArea] type.
+enum DockingAreaType { item, tabs, row, column }