+++ test/find_test.dart
@@ -15,30 +15,33 @@ void main() {
     test('row', () {
       DockingItem item1 = dockingItem(1);
       DockingItem item2 = dockingItem(2);
-      DockingRow row = DockingRow([item1, item2]);
+      DockingRow row = DockingRow([item1, item2], id: 'row');
       DockingLayout layout = DockingLayout(root: row);
-      expect(layout.findDockingItem(1), isNotNull);
-      expect(layout.findDockingItem(2), isNotNull);
+      expect(layout.findDockingItem(1), item1);
+      expect(layout.findDockingItem(2), item2);
       expect(layout.findDockingItem(3), isNull);
+      expect(layout.findDockingArea('row'), row);
     });
     test('column', () {
       DockingItem item1 = dockingItem(1);
       DockingItem item2 = dockingItem(2);
-      DockingColumn column = DockingColumn([item1, item2]);
+      DockingColumn column = DockingColumn([item1, item2], id: 'column');
       DockingLayout layout = DockingLayout(root: column);
-      expect(layout.findDockingItem(1), isNotNull);
-      expect(layout.findDockingItem(2), isNotNull);
+      expect(layout.findDockingItem(1), item1);
+      expect(layout.findDockingItem(2), item2);
       expect(layout.findDockingItem(3), isNull);
+      expect(layout.findDockingArea('column'), column);
     });
 
     test('tabs', () {
-      DockingItem itemB = dockingItem(1);
-      DockingItem itemC = dockingItem(2);
-      DockingTabs tabs = DockingTabs([itemB, itemC]);
+      DockingItem item1 = dockingItem(1);
+      DockingItem item2 = dockingItem(2);
+      DockingTabs tabs = DockingTabs([item1, item2], id: 'tabs');
       DockingLayout layout = DockingLayout(root: tabs);
-      expect(layout.findDockingItem(1), isNotNull);
-      expect(layout.findDockingItem(2), isNotNull);
+      expect(layout.findDockingItem(1), item1);
+      expect(layout.findDockingItem(2), item2);
       expect(layout.findDockingItem(3), isNull);
+      expect(layout.findDockingArea('tabs'), tabs);
     });
 
     test('complex', () {
@@ -47,18 +50,23 @@ void main() {
       DockingItem itemC = dockingItem('c');
       DockingItem itemD = dockingItem('d');
       DockingItem itemE = dockingItem('e');
-      DockingColumn innerColumn = DockingColumn([itemB, itemC]);
-      DockingRow row = DockingRow([itemA, innerColumn]);
-      DockingTabs tabs = DockingTabs([itemD, itemE]);
-      DockingColumn column = DockingColumn([row, tabs]);
+      DockingColumn innerColumn =
+          DockingColumn([itemB, itemC], id: 'innerColumn');
+      DockingRow row = DockingRow([itemA, innerColumn], id: 'row');
+      DockingTabs tabs = DockingTabs([itemD, itemE], id: 'tabs');
+      DockingColumn column = DockingColumn([row, tabs], id: 'column');
       DockingLayout layout = DockingLayout(root: column);
 
-      expect(layout.findDockingItem('a'), isNotNull);
-      expect(layout.findDockingItem('b'), isNotNull);
-      expect(layout.findDockingItem('c'), isNotNull);
-      expect(layout.findDockingItem('d'), isNotNull);
-      expect(layout.findDockingItem('e'), isNotNull);
+      expect(layout.findDockingItem('a'), itemA);
+      expect(layout.findDockingItem('b'), itemB);
+      expect(layout.findDockingItem('c'), itemC);
+      expect(layout.findDockingItem('d'), itemD);
+      expect(layout.findDockingItem('e'), itemE);
       expect(layout.findDockingItem('f'), isNull);
+      expect(layout.findDockingArea('innerColumn'), innerColumn);
+      expect(layout.findDockingArea('row'), row);
+      expect(layout.findDockingArea('column'), column);
+      expect(layout.findDockingArea('innerColumn'), innerColumn);
     });
 
     test('complex 2', () {
@@ -68,19 +76,24 @@ void main() {
       DockingItem itemD = dockingItem('d');
       DockingItem itemE = dockingItem('e');
       DockingItem itemF = dockingItem('f');
-      DockingColumn innerColumn = DockingColumn([itemB, itemC]);
-      DockingRow row = DockingRow([itemA, innerColumn]);
-      DockingTabs tabs = DockingTabs([itemD, itemE]);
-      DockingColumn column = DockingColumn([row, tabs, itemF]);
+      DockingColumn innerColumn =
+          DockingColumn([itemB, itemC], id: 'innerColumn');
+      DockingRow row = DockingRow([itemA, innerColumn], id: 'row');
+      DockingTabs tabs = DockingTabs([itemD, itemE], id: 'tabs');
+      DockingColumn column = DockingColumn([row, tabs, itemF], id: 'column');
       DockingLayout layout = DockingLayout(root: column);
 
-      expect(layout.findDockingItem('a'), isNotNull);
-      expect(layout.findDockingItem('b'), isNotNull);
-      expect(layout.findDockingItem('c'), isNotNull);
-      expect(layout.findDockingItem('d'), isNotNull);
-      expect(layout.findDockingItem('e'), isNotNull);
-      expect(layout.findDockingItem('f'), isNotNull);
+      expect(layout.findDockingItem('a'), itemA);
+      expect(layout.findDockingItem('b'), itemB);
+      expect(layout.findDockingItem('c'), itemC);
+      expect(layout.findDockingItem('d'), itemD);
+      expect(layout.findDockingItem('e'), itemE);
+      expect(layout.findDockingItem('f'), itemF);
       expect(layout.findDockingItem('g'), isNull);
+      expect(layout.findDockingArea('innerColumn'), innerColumn);
+      expect(layout.findDockingArea('row'), row);
+      expect(layout.findDockingArea('column'), column);
+      expect(layout.findDockingArea('innerColumn'), innerColumn);
     });
 
     test('complex 3', () {
@@ -91,21 +104,27 @@ void main() {
       DockingItem itemE = dockingItem('e');
       DockingItem itemF = dockingItem('f');
       DockingItem itemG = dockingItem('g');
-      DockingColumn innerColumn = DockingColumn([itemB, itemC]);
-      DockingRow row = DockingRow([itemA, innerColumn]);
-      DockingTabs tabs = DockingTabs([itemD, itemE]);
-      DockingColumn column = DockingColumn([row, tabs, itemF]);
-      DockingRow row2 = DockingRow([itemG, column]);
+      DockingColumn innerColumn =
+          DockingColumn([itemB, itemC], id: 'innerColumn');
+      DockingRow row = DockingRow([itemA, innerColumn], id: 'row');
+      DockingTabs tabs = DockingTabs([itemD, itemE], id: 'tabs');
+      DockingColumn column = DockingColumn([row, tabs, itemF], id: 'column');
+      DockingRow row2 = DockingRow([itemG, column], id: 'row2');
       DockingLayout layout = DockingLayout(root: row2);
 
-      expect(layout.findDockingItem('a'), isNotNull);
-      expect(layout.findDockingItem('b'), isNotNull);
-      expect(layout.findDockingItem('c'), isNotNull);
-      expect(layout.findDockingItem('d'), isNotNull);
-      expect(layout.findDockingItem('e'), isNotNull);
-      expect(layout.findDockingItem('f'), isNotNull);
-      expect(layout.findDockingItem('g'), isNotNull);
+      expect(layout.findDockingItem('a'), itemA);
+      expect(layout.findDockingItem('b'), itemB);
+      expect(layout.findDockingItem('c'), itemC);
+      expect(layout.findDockingItem('d'), itemD);
+      expect(layout.findDockingItem('e'), itemE);
+      expect(layout.findDockingItem('f'), itemF);
+      expect(layout.findDockingItem('g'), itemG);
       expect(layout.findDockingItem('h'), isNull);
+      expect(layout.findDockingArea('innerColumn'), innerColumn);
+      expect(layout.findDockingArea('row'), row);
+      expect(layout.findDockingArea('row2'), row2);
+      expect(layout.findDockingArea('column'), column);
+      expect(layout.findDockingArea('innerColumn'), innerColumn);
     });
   });
 }